feat(DTFS2-7653): improve deal, deal cancellation services
