+++ dtfs-central-api/src/services/portal/deal.service.ts
@@ -1,4 +1,4 @@
-import { AuditDetails, ActivityAuthor, DEAL_TYPE, DealStatus, DealType, PORTAL_ACTIVITY_LABEL, TfmDeal, UKEF } from '@ukef/dtfs2-common';
+import { AuditDetails, ActivityAuthor, DEAL_TYPE, DealStatus, DealType, PORTAL_ACTIVITY_LABEL, UKEF } from '@ukef/dtfs2-common';
 import { getUnixTime } from 'date-fns';
 import { ObjectId } from 'mongodb';
 import { updateBssEwcsDealStatus } from '../../v1/controllers/portal/deal/update-deal-status.controller';
@@ -43,31 +43,60 @@ export class PortalDealService {
   }
 
   /**
+   * Add a "GEF deal cancellation pending" activity
+   * @param addGefDealCancellationPendingActivity
+   * @param addGefDealCancellationPendingActivity.dealId - the deal ID
+   * @param addGefDealCancellationPendingActivity.dealType - the deal type
+   * @param addGefDealCancellationPendingActivity.author - the activity's author
+   * @param addGefDealCancellationPendingActivity.auditDetails - the users audit details
+   */
+  public static async addGefDealCancellationPendingActivity({
+    dealId,
+    dealType,
+    author,
+    auditDetails,
+  }: {
+    dealId: ObjectId | string;
+    dealType: DealType;
+    author: ActivityAuthor;
+    auditDetails: AuditDetails;
+  }): Promise<void> {
+    if (dealType === DEAL_TYPE.GEF) {
+      const newActivity = {
+        label: PORTAL_ACTIVITY_LABEL.DEAL_CANCELLATION_PENDING,
+        timestamp: getUnixTime(new Date()),
+        author: {
+          _id: author._id,
+          firstName: UKEF.ACRONYM,
+        },
+      };
+
+      await PortalActivityRepo.addPortalActivity(dealId, newActivity, auditDetails);
+    }
+  }
+
+  /**
    * Add a "GEF deal cancelled" activity
    * @param addGefDealCancelledActivityParams
-   * @param addGefDealCancelledActivityParams.dealId - the deal
+   * @param addGefDealCancelledActivityParams.dealId - the deal ID
+   * @param addGefDealCancelledActivityParams.dealType - the deal type
    * @param addGefDealCancelledActivityParams.author - the activity's author
-   * @param addGefDealCancelledActivityParams.cancellationIsInFuture - cancellation is in the future
    * @param addGefDealCancelledActivityParams.auditDetails - the users audit details
    */
   public static async addGefDealCancelledActivity({
-    deal,
+    dealId,
+    dealType,
     author,
-    cancellationIsInFuture,
     auditDetails,
   }: {
-    deal: TfmDeal;
+    dealId: ObjectId | string;
+    dealType: DealType;
     author: ActivityAuthor;
-    cancellationIsInFuture?: boolean;
     auditDetails: AuditDetails;
   }): Promise<void> {
-    if (deal.dealSnapshot.dealType === DEAL_TYPE.GEF) {
-      const { _id: dealId } = deal.dealSnapshot;
-
-      const label = cancellationIsInFuture ? PORTAL_ACTIVITY_LABEL.DEAL_CANCELLATION_PENDING : PORTAL_ACTIVITY_LABEL.DEAL_CANCELLED;
-
+    if (dealType === DEAL_TYPE.GEF) {
       const newActivity = {
-        label,
+        label: PORTAL_ACTIVITY_LABEL.DEAL_CANCELLED,
         timestamp: getUnixTime(new Date()),
         author: {
           _id: author._id,
